add juju availability zone to ceph osd location when present
The approach here is to use the availability zone as an imaginary rack.
All hosts that are in the same AZ will be in the same imaginary rack.
From Ceph's perspective this doesn't matter as it's just a bucket after all.
This will give users the ability to further customize their ceph deployment.

Change-Id: Ie25ac1b001db558d6a40fe3eaca014e8f4174241