Upgrade OSDs one at a time when changing ownership
Some upgrade scenarios (hammer->jewel) require that the ownership
of the ceph osd directories are changed from root:root to ceph:ceph.
This patch improves the upgrade experience by upgrading one OSD at
a time as opposed to stopping all services, changing file ownership,
and then restarting all services at once.

This patch makes use of the `setuser match path` directive in the
ceph.conf, which causes the ceph daemon to start as the owner of the
OSD's root directory. This allows the ceph OSDs to continue running
should an unforeseen incident occur as part of this upgrade.

Change-Id: I00fdbe0fd113c56209429341f0a10797e5baee5a
Closes-Bug: #1662591